requirements.txt
# Core
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7

# Crawler
requests==2.32.4
beautifulsoup4==4.13.4

# Embeddings & vector search (lightweight – runs on CPU)
sentence-transformers==5.0.0
faiss-cpu==1.11.0.post1
numpy==1.26.4
scikit-learn==1.7.0

# LLM via cloud API
groq==0.25.0

# Templating (Jinja used by FastAPI)
Jinja2==3.1.6
